refactor(shortcuts): move help directory out of components/commands
The keyboard-shortcuts help overlay + shared list use the standard white/neutral
modal styling (ModalShell), which the commands/ directory's cmd-token lint
(lint:commands-tokens) forbids. Relocate shortcutsView + KeyboardShortcutsModal
(and their test) to components/shortcuts/ and repoint imports; Kbd stays in
commands/ and is imported across.
