+++ app/src/components/shortcuts/__tests__/shortcutsView.test.tsx
